auth_service/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
pymongo>=4.8,<5.0
passlib[bcrypt]>=1.7.4,<2.0
python-jose[cryptography]>=3.3,<4.0
python-multipart>=0.0.9,<1.0
email-validator>=2.2,<3.0
